# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI-HRMS API
  version: 0.1.0
  description: >
    Contract index for the current AI-HRMS reproduction phase. The API serves
    one corporate group with multiple legal entities; legal entities are
    organization scopes, not SaaS tenants.
servers:
  - url: http://localhost:8080/api
    description: Local Go API
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login with mobile and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token and current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /profile:
    get:
      tags: [Auth]
      operationId: getProfile
      summary: Get current user profile
      responses:
        "200":
          description: Current authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /legal-entities:
    get:
      tags: [Organization]
      operationId: listLegalEntities
      summary: List visible legal entities
      description: Returns legal entities visible to the caller's RBAC scope.
      responses:
        "200":
          description: Visible legal entities.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalEntityListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Organization]
      operationId: createLegalEntity
      summary: Create legal entity
      description: Requires global role scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalEntityInput"
      responses:
        "201":
          description: Created legal entity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalEntityEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /legal-entities/{id}:
    put:
      tags: [Organization]
      operationId: updateLegalEntity
      summary: Update legal entity
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegalEntityInput"
      responses:
        "200":
          description: Updated legal entity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalEntityEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /org-units:
    get:
      tags: [Organization]
      operationId: listOrgUnits
      summary: List visible organization units
      description: Organization units may have no legal entity.
      responses:
        "200":
          description: Visible organization units.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnitListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Organization]
      operationId: createOrgUnit
      summary: Create organization unit
      description: Requires global role scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgUnitInput"
      responses:
        "201":
          description: Created organization unit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnitEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /org-units/{id}:
    put:
      tags: [Organization]
      operationId: updateOrgUnit
      summary: Update organization unit
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgUnitInput"
      responses:
        "200":
          description: Updated organization unit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgUnitEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /roles:
    get:
      tags: [Roles]
      operationId: listRoles
      summary: List roles
      description: Requires global role scope.
      responses:
        "200":
          description: Available roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users:
    get:
      tags: [Users]
      operationId: listUsers
      summary: List users
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Users]
      operationId: createUser
      summary: Create user
      description: Requires global role scope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: Created user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /users/{id}:
    put:
      tags: [Users]
      operationId: updateUser
      summary: Update user
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{id}/role-bindings:
    get:
      tags: [Roles]
      operationId: listUserRoleBindings
      summary: List user role bindings
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: User role bindings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBindingListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Roles]
      operationId: replaceUserRoleBindings
      summary: Replace user role bindings
      description: Requires global role scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceRoleBindingsRequest"
      responses:
        "200":
          description: Updated user role bindings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleBindingListEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /employees:
    get:
      tags: [Employees]
      operationId: listEmployees
      summary: List employees with primary assignment
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged employees visible to caller scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Employees]
      operationId: createEmployee
      summary: Create employee
      description: Requires global role scope in the current implementation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /employees/export:
    get:
      tags: [Employees]
      operationId: exportEmployees
      summary: Export visible employees as CSV
      responses:
        "200":
          description: UTF-8 CSV file.
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /employees/{id}:
    get:
      tags: [Employees]
      operationId: getEmployee
      summary: Get employee by id
      description: Returns only employees visible to caller scope.
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Employee detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Employees]
      operationId: updateEmployee
      summary: Update employee
      description: Requires global role scope in the current implementation.
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Updated employee.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /attendance:
    get:
      tags: [Attendance]
      operationId: listAttendance
      summary: List attendance records
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged attendance visible to caller scope.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendancePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Attendance]
      operationId: createAttendance
      summary: Check in an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceRequest"
      responses:
        "201":
          description: Created attendance record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /attendance/export:
    get:
      tags: [Attendance]
      operationId: exportAttendance
      summary: Export visible attendance records as CSV
      responses:
        "200":
          description: UTF-8 CSV file.
          content:
            text/csv:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
  /attendance/{id}/checkout:
    put:
      tags: [Attendance]
      operationId: checkoutAttendance
      summary: Check out an attendance record
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Updated attendance record.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /messages:
    get:
      tags: [Messages]
      operationId: listMessages
      summary: List messages
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: Paged messages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePageEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Messages]
      operationId: createMessage
      summary: Create message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "201":
          description: Created message.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /messages/{id}/comments:
    get:
      tags: [Messages]
      operationId: listComments
      summary: List message comments
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Message comments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListEnvelope"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Messages]
      operationId: createComment
      summary: Create message comment
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
      responses:
        "201":
          description: Created comment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentEnvelope"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /rag/ingest-jobs/{id}:
    get:
      tags: [RAG]
      operationId: getRagIngestJob
      summary: Get ingestion job
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Ingestion job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /learning/courses/{id}/lessons:
    get:
      tags: [Learning]
      operationId: listLearningLessons
      summary: List lessons for a course
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lesson list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing, invalid, or expired bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Caller lacks the required RBAC scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Resource not found or outside caller scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    LoginEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/LoginResult"
    UserEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/User"
    LegalEntityEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/LegalEntity"
    LegalEntityListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/LegalEntity"
    OrgUnitEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/OrgUnit"
    OrgUnitListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/OrgUnit"
    RoleListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Role"
    RoleBindingListEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/RoleBinding"
    EmployeeEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/Employee"
...
    UserPageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/UserPage"
    EmployeePageEnvelope:
      allOf:
        - $ref: "#/components/schemas/ApiEnvelope"
        - type: object
          required: [data]
          properties:
            data:
              $ref: "#/components/schemas/EmployeePage"
...
    LoginRequest:
      type: object
      required: [mobile, password]
      properties:
        mobile:
          type: string
          example: "123"
        password:
          type: string
          format: password
          example: password
...
    CreateUserRequest:
      type: object
      required: [mobile, username]
      properties:
        mobile:
          type: string
        username:
          type: string
        password:
          type: string
          format: password
          default: password
    UpdateUserRequest:
      type: object
      required: [username, enableState]
      properties:
        username:
          type: string
        enableState:
          type: integer
          enum: [0, 1]
    ReplaceRoleBindingsRequest:
      type: object
      required: [bindings]
      properties:
        bindings:
          type: array
          items:
            $ref: "#/components/schemas/RoleBindingInput"
...
    LegalEntityInput:
      type: object
      required: [code, name]
      properties:
        parentId:
          type: [string, "null"]
          format: uuid
        code:
          type: string
        name:
          type: string
        legalName:
          type: string
        unifiedSocialCreditCode:
          type: string
        legalRepresentative:
          type: string
        companyPhone:
          type: string
        email:
          type: string
          format: email
        area:
          type: string
        address:
          type: string
        status:
          type: string
          default: active
    OrgUnitInput:
      type: object
      required: [code, name]
      properties:
        parentId:
          type: [string, "null"]
          format: uuid
        legalEntityId:
          type: [string, "null"]
          format: uuid
        code:
          type: string
        name:
          type: string
        type:
          type: string
          default: department
        managerName:
          type: string
        status:
          type: string
          default: active
...